Library/fips: error on import if misconfigured
`fipsIsEnabled` returns 0 on 'enabled', 1 on 'disabled' or 2 on
'misconfigured'. This does not really align with common usage, i.e.,
using `fipsIsEnabled` and `! fipsIsEnabled` without checking for that 2.

I'm not aware of the test cases that use `fipsIsEnabled` to verify that
FIPS mode is *succesfully misconfigured*, but there may be some, so
making `fipsIsEnabled` `rlFail` on the misconfiguration detection might
be a tad too invasive.

On the other hand, I'm fairly sure that having FIPS broken on
`rlImport`-time is not something that is ever desirable,
even in tests that have to reboot mid-execution.

This change has two consequences:
* FIPS misconfiguration at import time results in `rlFail`
  (for the common case it will flip setupPhase to WARN and go on)
* A log line on whether FIPS is enabled appear on `rlImport`
  (doesn't sound bad, actually)