Make restraint workaround more robust
On system with restraint packages are installed and removed using
custom wrapper. There was a minor issue cause that this workaround
used the same temporary directory for every invocation. That is not
always bad but the original intention was to use different directory
every time. Moreover, we are sligthly extending self-test.